Splitting up the three signal classes to its own files
