+++ src/BlazorLighthouse/Core/ReadonlySignal.cs
@@ -1,0 +1,20 @@
+﻿namespace BlazorLighthouse.Core;
+
+/// <summary>
+/// Simple readonly value store that allows subscription to changes
+/// </summary>
+/// <typeparam name="T">Value type</typeparam>
+public abstract class ReadonlySignal<T> : SignalBase
+{
+    internal ReadonlySignal(SignalingContext context)
+        : base(context)
+    {
+
+    }
+
+    /// <summary>
+    /// Get the current value stored and registers available subscribers to changes
+    /// </summary>
+    /// <returns>The current value stored</returns>
+    public abstract T Get();
+}
