+++ src/BlazorLighthouse/Core/Signal.cs
@@ -1,88 +1,6 @@
 ﻿using BlazorLighthouse.Internal;
-using BlazorLighthouse.Internal.Interfaces;
 
 namespace BlazorLighthouse.Core;
-
-/// <summary>
-/// Base class for all types of signals
-/// </summary>
-public abstract class Signal : IContextDisposable
-{
-    /// <summary>
-    /// Signaling context for the current signal
-    /// </summary>
-    protected readonly SignalingContext context;
-
-    private HashSet<IRefreshable> refreshables = [];
-
-    internal Signal(SignalingContext context)
-    {
-        this.context = context;
-        context?.RegisterContextDisposable(this);
-    }
-
-    /// <summary>
-    /// Notify the signaling system that the value has chagned
-    /// </summary>
-    internal protected void ValueHasChanged()
-    {
-        var currentRefreshables = refreshables;
-        refreshables = [];
-
-        Refresh(currentRefreshables);
-    }
-
-    internal void RegisterRefreshable(IRefreshable refreshable)
-    {
-        lock (context.LockObject)
-        {
-            RegisterRefreshableSynchronized(refreshable);
-        }
-    }
-
-    internal void UnregisterRefreshable(IRefreshable refreshable)
-    {
-        refreshables.Remove(refreshable);
-    }
-    
-    private void RegisterRefreshableSynchronized(IRefreshable refreshable)
-    {
-        context.AssertIsNotDisposed();
-        refreshables.Add(refreshable);
-    }
-
-    void IContextDisposable.Dispose()
-    {
-        refreshables.Clear();
-        foreach (var refreshable in refreshables)
-            refreshable.Dispose(this);
-    }
-
-    private static void Refresh(ISet<IRefreshable> refreshables)
-    {
-        foreach (var refreshable in refreshables)
-            refreshable.Refresh();
-    }
-}
-
-/// <summary>
-/// Simple readonly value store that allows subscription to changes
-/// </summary>
-/// <typeparam name="T">Value type</typeparam>
-public abstract class ReadonlySignal<T> : Signal
-{
-    internal ReadonlySignal(SignalingContext context)
-        : base(context)
-    {
-
-    }
-
-    /// <summary>
-    /// Get the current value stored and registers available subscribers to changes
-    /// </summary>
-    /// <returns>The current value stored</returns>
-    public abstract T Get();
-}
 
 /// <summary>
 /// Simple value store that allows subscription to changes
